third_party/WebKit/LayoutTests/platform/win/security/cannot-read-self-from-file-expected.txt

CONSOLE ERROR: line 10: XMLHttpRequest cannot load cannot-read-self-from-file.html. Cross origin requests are only supported for protocol schemes: http, data, chrome, https.
This is a testharness.js-based test.
PASS file: should be a unique-origin protocol for XHR purposes 
Harness: the test ran to completion.
